feat: #15, Modal ContextAPI를 zustand로 마이그레이션
