+++ src/@types/modal.type.ts
@@ -1,8 +1,8 @@
 import { ReactNode } from 'react';
 
+export const PORTAL_ELEMENT_ID = 'modal';
 export type Modal = {
+  element: ReactNode | null;
   showModal: (children: ReactNode) => void;
   closeModal: () => void;
-  isVisible: boolean;
 };
-export const PORTAL_ELEMENT_ID = 'modal';
